refactor: rename and relocate getRawImages controller to getRawImage
